IBOpsSim/webapp/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyyaml==6.0.1
simpy==4.0.2
setuptools>=65.0.0
